Initial Push
I have added functionality to draw edges between graph nodes based on the edges defined in 'edge_config.md.' If edges become invalid due to updates in that file or if waypoints are removed from the map, these edges will disappear. Additionally, programmatic markers (designated by the #programmatic-markers tag) will be sized based on the degree of a given vertex.